+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.22.0] - 2022-07-16
+### Added
+- `capitalize` method
+### Fixed
+- `throttle` reset not resetting array of times
+
 ## [0.21.0] - 2022-07-08
 ### Added
 - `throttle` method